Fix error in requires_start_end decorator
The requires_start_end decorator was incorrectly raising an error when either the start or end attribute of the clip was not set. This has been fixed to only raise an error when both attributes are not set. Additionally, the error message now includes the value of the end attribute.