Add set_path function to config.py
This commit adds a new function called set_path to the config.py file. The set_path function allows users to set the paths for the ffmpeg and ffprobe binaries. This function takes two optional parameters, ffmpeg_path and ffprobe_path, which can be used to specify the paths for the binaries. If no paths are provided, the function will use the default paths. This new function provides more flexibility for users to customize the paths for the ffmpeg and ffprobe binaries.